Fix EDITOR and GIT_EDITOR being unset initially
Due to the order of files in the sourcing array, VISUAL was being set
after EDITOR and GIT_EDITOR, which meant that the latter two were empty
variables. On initial system boot, things reliant on those variables
would fail to open editors.

As such, I've added finally.sh, which will have exports or other config
that can only be done after all other aliases and exports have been set.
This is admittedly a bit of a hack, but I think it's a cleaner method
than changing the insertion order of the other source files.